Extract CategorizationOrchestrator collaborator
Promotes the batch-categorization flows to a first-class concept in a
new categorization/orchestrator.py. CategorizationOrchestrator composes a
CategorizationMatcher and a MatchApplier and owns the five flows that
drive transactions through matching to produce category writes:

- categorize_items / _categorize_items_inner: apply an LLM-validated
  batch with merchant resolution, auto-rule observation, exemplar
  accumulation, and a post-commit snowball.
- apply_rules: sweep uncategorized rows against active categorization
  rules.
- apply_merchant_categories: sweep uncategorized rows against the
  merchant catalog.
- categorize_pending: combined snowball (rules pass then merchants
  pass over a single shared scan).

CategorizationResult lives in orchestrator.py with the methods that
produce it. The lazy import of AutoRuleService stays — it preserves the
one-way module dependency (auto_rule_service → categorization).

Facade __init__.py drops from 944 to 444 lines — within striking distance
of the plan's ~400 target. Every collaborator file is under the 800-line
DoD limit:

  __init__.py    444   (facade)
  applier.py     722
  orchestrator.py 634
  matcher.py     391
  _shared.py     331
  queries.py     304
  assist.py      145

No behavior change; full suite passes.
